BOCM-486 - moves endpoints routing to own file
